requirements.txt
# place any necessary python libraries here, one per line, to be installed with pip install -r requirements.txt. See https://pip.readthedocs.io/en/1.1/requirements.html

configparser
TwitterAPI
networkx
numpy
re
collections
sys
time
pickle
csv
matplotlib.pyplot
math
itertools